Change the DELETE route

+++ backend/controllers/sauce.js
@@ -68,18 +68,25 @@ exports.modify = (req, res) => {
       });
 };
 
-
-exports.delete = (req, res) => {
-Sauce.deleteOne({ _id: req.params.id })
-  .then(() => {
-    res.status(200).json({ message: 'Objet supprimé !'})
-    console.log('DELETE request successful');
-  })
-  .catch(error => {
-    res.status(400).json({ error })
-    console.log('DELETE request failed. Error:', error);
-  });
+exports.delete = (req, res, next) => {
+  Sauce.findOne({ _id: req.params.id})
+      .then(sauce => {
+          if (sauce.userId != req.auth.userId) {
+              res.status(401).json({message: 'Non autorisé'});
+          } else {
+              const filename = sauce.imageUrl.split('/images/')[1];
+              fs.unlink(`images/${filename}`, () => {
+                  Sauce.deleteOne({_id: req.params.id})
+                      .then(() => { res.status(200).json({message: 'Sauce supprimé !'})})
+                      .catch(error => res.status(401).json({ error }));
+              });
+          }
+      })
+      .catch( error => {
+          res.status(500).json({ error });
+      });
 };
+
 
 exports.like = (req , res) => {
